Fix static analysis warning about possible overflow #preflight 63d0e1b2033e457ba8a20812
[CL 23845558 by andrean franc in ue5-main branch]

+++ Engine/Plugins/Animation/RigLogic/Source/RigLogicLib/Private/riglogic/riglogic/RigInstanceImpl.cpp
@@ -151,7 +151,7 @@ float RigInstanceImpl::getMLControl(std::uint16_t index) const {
 }
 
 ConstArrayView<float> RigInstanceImpl::getMLControlValues() const {
-    const auto mlControlsOffset = static_cast<std::size_t>(rawControlCount + psdControlCount);
+    const auto mlControlsOffset = static_cast<std::size_t>(rawControlCount) + static_cast<std::size_t>(psdControlCount);
     return controlsInstance->getInputBuffer().subview(mlControlsOffset, mlControlCount);
 }
